requirements.txt
streamlit==0.81.1
pandas==1.3.1
numpy==1.21.1
matplotlib==3.4.3
seaborn==0.11.2
scikit-learn==0.24.2
scipy==1.7.1
klib==0.2.0
pillow==8.3.1
cython>= 0.28.5
xgboost==1.4.2




